extract route enhancer into trait
