+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Changelog
 =========
 
+ * **2017-01-31**: Split out enhancer code from DynamicRouter into RouteEnhancerTrait for reusability.
+
 2.0.0-RC1
 ---------
